X86_64: silence some truncation warnings
There are a number of places where we truncate to a 32-bit value from a
64-bit value as we may running a 32-bit binary on 64-bits. This silences
some of the warnings by explicitly truncating the integers rather than
implicitly performing the operation.

+++ Headers/DebugServer2/Core/HardwareBreakpointManager.h
@@ -75,7 +75,7 @@ protected:
 protected:
   virtual ErrorCode disableDebugCtrlReg(uint64_t &ctrlReg, int idx);
   virtual ErrorCode enableDebugCtrlReg(uint64_t &ctrlReg, int idx, Mode mode,
-                                       int size);
+                                       size_t size);
 #endif
 
 public:
